+++ packages/client-connectivity/src/lib/security.ts
@@ -177,7 +177,10 @@ export function resolveGlobalSecurity(
         [
             {
                 type: "oauth2:client_credentials",
-                value: security?.formanceOAuth,
+                value: {
+                    clientID: security?.formanceOAuth?.clientID,
+                    clientSecret: security?.formanceOAuth?.clientSecret,
+                },
             },
         ],
         [
@@ -188,4 +191,14 @@ export function resolveGlobalSecurity(
             },
         ]
     );
+}
+
+export async function extractSecurity<T extends string | Record<string, unknown>>(
+    sec: T | (() => Promise<T>) | undefined
+): Promise<T | undefined> {
+    if (sec == null) {
+        return;
+    }
+
+    return typeof sec === "function" ? sec() : sec;
 }