+++ packages/client-connectivity/src/sdk/transferinitiations.ts
@@ -13,6 +13,7 @@ import {
 import { HTTPClient } from "../lib/http.js";
 import * as schemas$ from "../lib/schemas.js";
 import { ClientSDK, RequestOptions } from "../lib/sdks.js";
+import { extractSecurity } from "../lib/security.js";
 import * as errors from "../models/errors/index.js";
 import * as operations from "../models/operations/index.js";
 import { createPageIterator, PageIterator, Paginator } from "../types/operations.js";
@@ -79,11 +80,7 @@ export class TransferInitiations extends ClientSDK {
             Accept: "application/json",
         });
 
-        const security$ =
-            typeof this.options$.security === "function"
-                ? await this.options$.security()
-                : this.options$.security;
-
+        const security$ = await extractSecurity(this.options$.security);
         const context = {
             operationID: "TransferInitiations_list",
             oAuth2Scopes: ["payments:read"],
@@ -174,11 +171,7 @@ export class TransferInitiations extends ClientSDK {
             Accept: "application/json",
         });
 
-        const security$ =
-            typeof this.options$.security === "function"
-                ? await this.options$.security()
-                : this.options$.security;
-
+        const security$ = await extractSecurity(this.options$.security);
         const context = {
             operationID: "TransferInitiations_create",
             oAuth2Scopes: ["payments:write"],
@@ -260,11 +253,7 @@ export class TransferInitiations extends ClientSDK {
             Accept: "application/json",
         });
 
-        const security$ =
-            typeof this.options$.security === "function"
-                ? await this.options$.security()
-                : this.options$.security;
-
+        const security$ = await extractSecurity(this.options$.security);
         const context = {
             operationID: "TransferInitiations_get",
             oAuth2Scopes: ["payments:read"],
@@ -343,11 +332,7 @@ export class TransferInitiations extends ClientSDK {
             Accept: "application/json",
         });
 
-        const security$ =
-            typeof this.options$.security === "function"
-                ? await this.options$.security()
-                : this.options$.security;
-
+        const security$ = await extractSecurity(this.options$.security);
         const context = {
             operationID: "TransferInitiations_delete",
             oAuth2Scopes: ["payments:write"],
@@ -426,11 +411,7 @@ export class TransferInitiations extends ClientSDK {
             Accept: "application/json",
         });
 
-        const security$ =
-            typeof this.options$.security === "function"
-                ? await this.options$.security()
-                : this.options$.security;
-
+        const security$ = await extractSecurity(this.options$.security);
         const context = {
             operationID: "TransferInitiations_retry",
             oAuth2Scopes: ["payments:write"],
@@ -515,11 +496,7 @@ export class TransferInitiations extends ClientSDK {
             Accept: "application/json",
         });
 
-        const security$ =
-            typeof this.options$.security === "function"
-                ? await this.options$.security()
-                : this.options$.security;
-
+        const security$ = await extractSecurity(this.options$.security);
         const context = {
             operationID: "TransferInitiations_reverse",
             oAuth2Scopes: ["payments:write"],
@@ -605,11 +582,7 @@ export class TransferInitiations extends ClientSDK {
             Accept: "application/json",
         });
 
-        const security$ =
-            typeof this.options$.security === "function"
-                ? await this.options$.security()
-                : this.options$.security;
-
+        const security$ = await extractSecurity(this.options$.security);
         const context = {
             operationID: "TransferInitiations_updateStatus",
             oAuth2Scopes: ["payments:write"],